[refactor] : Store 테이블의 온도 컬럼은 기본 값이 36.5 로 정해져 있고, null 값을 가지지 않으므로 원시 타입으로 변경

+++ server/src/main/java/nior_near/server/domain/store/dto/response/StoreResponseDto.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 import nior_near.server.domain.store.entity.Menu;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -21,7 +20,7 @@ public class StoreResponseDto {
     Long placeId;
     String placeName;
     List<String> auths;
-    BigDecimal temperature;
+    double temperature;
     List<MenuItem> menus;
 
     @Getter
